# artifacts/openapi.json
{
  "components": {
    "schemas": {
      "BatchCreateRequest": {
        "examples": [
          {
            "description": "Example batch for the frontend contract.",
            "name": "Batch 1"
          }
        ],
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "BatchCreateRequest",
        "type": "object"
      },
      "BatchResponse": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/BatchStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "status",
          "created_by",
          "created_at",
          "updated_at",
          "started_at",
          "completed_at"
        ],
        "title": "BatchResponse",
        "type": "object"
      },
      "BatchStatus": {
        "enum": [
          "DRAFT",
          "QUEUED",
          "PROCESSING",
          "REVIEW_REQUIRED",
          "COMPLETED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "BatchStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OneDriveLinkCreateRequest": {
        "examples": [
          {
            "url": "https://1drv.ms/f/s!AbCdEfGhIjKlMnOp"
          }
        ],
        "properties": {
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "OneDriveLinkCreateRequest",
        "type": "object"
      },
      "OneDriveSubmissionError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "OneDriveSubmissionError",
        "type": "object"
      },
      "OneDriveSubmissionResponse": {
        "properties": {
          "batch_id": {
            "format": "uuid",
            "title": "Batch Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OneDriveSubmissionError"
              },
              {
                "type": "null"
              }
            ]
          },
          "fetched_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "skipped_files": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SkippedFile"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skipped Files"
          },
          "status": {
            "$ref": "#/components/schemas/OneDriveSubmissionStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "batch_id",
          "url",
          "status",
          "skipped_files",
          "created_at",
          "updated_at",
          "fetched_at"
        ],
        "title": "OneDriveSubmissionResponse",
        "type": "object"
      },
      "OneDriveSubmissionStatus": {
        "enum": [
          "PENDING",
          "FETCHING",
          "FETCHED",
          "FAILED"
        ],
        "title": "OneDriveSubmissionStatus",
        "type": "string"
      },
      "PaginatedBatches": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "limit",
          "offset",
          "total"
        ],
        "title": "PaginatedBatches",
        "type": "object"
      },
...
      "PaginatedOneDriveSubmissions": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OneDriveSubmissionResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "limit",
          "offset",
          "total"
        ],
        "title": "PaginatedOneDriveSubmissions",
        "type": "object"
      },
...
      "SkippedFile": {
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "filename",
          "status"
        ],
        "title": "SkippedFile",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "FastAPI backend for the Marriage OCR pipeline.",
    "title": "Marriage OCR API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/batches": {
      "get": {
        "operationId": "list_batches",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedBatches"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List All Batches",
        "tags": [
          "batches"
        ]
      },
      "post": {
        "operationId": "create_batch",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "batch": {
                  "summary": "Create a batch",
                  "value": {
                    "description": "Example batch for the frontend contract.",
                    "name": "Batch 1"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/BatchCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create One Batch",
        "tags": [
          "batches"
        ]
      }
    },
    "/api/v1/batches/{batch_id}": {
      "delete": {
        "operationId": "delete_batch",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete One Batch",
        "tags": [
          "batches"
        ]
      },
      "get": {
        "operationId": "get_batch",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get One Batch",
        "tags": [
          "batches"
        ]
      }
    },
...
    "/api/v1/batches/{batch_id}/onedrive-links": {
      "get": {
        "operationId": "list_onedrive_links",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedOneDriveSubmissions"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Onedrive Links",
        "tags": [
          "onedrive"
        ]
      },
      "post": {
        "operationId": "submit_onedrive_link",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "link": {
                  "summary": "Submit a OneDrive share link",
                  "value": {
                    "url": "https://1drv.ms/f/s!AbCdEfGhIjKlMnOp"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/OneDriveLinkCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OneDriveSubmissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Onedrive Link",
        "tags": [
          "onedrive"
        ]
      }
    },
    "/api/v1/batches/{batch_id}/onedrive-links/{submission_id}": {
      "delete": {
        "description": "Deletes one OneDrive link and everything it caused to be ingested --\nits documents, their OCR jobs, and any records those jobs produced.\nThe batch itself, and anything ingested by its other links, are left\nalone.",
        "operationId": "delete_onedrive_link",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Submission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Onedrive Link",
        "tags": [
          "onedrive"
        ]
      }
    },
    "/api/v1/batches/{batch_id}/onedrive-links/{submission_id}/retry": {
      "post": {
        "operationId": "retry_onedrive_link",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "submission_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Submission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OneDriveSubmissionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retry Onedrive Link",
        "tags": [
          "onedrive"
        ]
      }
    },
...
  },
...
}